+++ lib/presentation/notifiers/comments_page/comments_page_notifier.dart
@@ -2,10 +2,10 @@ import 'package:blogify_flutter_main/domain/data_sources/remote/comments_data_so
 import 'package:blogify_flutter_main/domain/entities/comment_entity.dart';
 import 'package:flutter/cupertino.dart';
 
-class GlobalMockCommentProvider extends ChangeNotifier {
+class CommentsPageProvider extends ChangeNotifier {
   final CommentsDataSource _commentsDataSource;
 
-  GlobalMockCommentProvider(this._commentsDataSource);
+  CommentsPageProvider(this._commentsDataSource);
 
   List<CommentEntity> get allComments => _commentsDataSource.allComments;
